+++ lib/vcr/internet_connection.rb
@@ -7,7 +7,8 @@ module VCR
     EXAMPLE_HOST = "example.com"
 
     def available?
-      @available ||= Ping.pingecho(EXAMPLE_HOST, 1, 80)
+      @available = Ping.pingecho(EXAMPLE_HOST, 1, 80) unless defined?(@available)
+      @available
     end
   end
 end