+++ spec/internet_connection_spec.rb
@@ -3,17 +3,35 @@ require 'spec_helper'
 describe VCR::InternetConnection do
   describe '.available?' do
     before(:each) do
-      described_class.instance_variable_set(:@available, nil)
+      described_class.send(:remove_instance_variable, :@available) if described_class.instance_variable_defined?(:@available)
     end
 
-    it 'returns true when pinging example.com succeeds' do
-      Ping.stub(:pingecho).with("example.com", anything, anything).and_return(true)
-      described_class.available?.should be_true
+    def stub_pingecho_with(value)
+      Ping.stub(:pingecho).with("example.com", anything, anything).and_return(value)
     end
 
-    it 'returns false when pinging example.com fails' do
-      Ping.stub(:pingecho).with("example.com", anything, anything).and_return(false)
-      described_class.available?.should be_false
+    context 'when pinging example.com succeeds' do
+      it 'returns true' do
+        stub_pingecho_with(true)
+        described_class.should be_available
+      end
+
+      it 'memoizes the value so no extra pings are made' do
+        Ping.should_receive(:pingecho).once.and_return(true)
+        3.times { described_class.available? }
+      end
+    end
+
+    context 'when pinging example.com fails' do
+      it 'returns false' do
+        stub_pingecho_with(false)
+        described_class.should_not be_available
+      end
+
+      it 'memoizes the value so no extra pings are made' do
+        Ping.should_receive(:pingecho).once.and_return(false)
+        3.times { described_class.available? }
+      end
     end
   end
 end